First cut of painting the player-owned edges.
This change is probably mostly OK except that I just made
sides public in game.rs and that ought to be done differently.

Also there is now almost duplicate code for determining the
player colour for this and for flows, which ought to be unified
in some nice way.

Other than that, this correctly paints the edges for each player.