Create Backend trait and implement in-memory version.
A Backend respresents a connection for a specific user to some
underlying game (whether that's hosted on a server, or locally in
memory). You can poll the Backend for additional actions that have been
performed, or submit an action to be sent back to the underlying game.

A GameView is a wrapper over top of any Backend that uses the action
stream from the Backend to create a hydrated game state. When the user
performs an action, it also performs the action on the local hydrated
game state and shows that to the user until it receives an update from
the server, so that server lag is not easy perceptible to the user.