Add UI for placing tiles.
Tiles are visible on the other player's board once placed, so they're
making it to the (in-memory) backend and back.

The UI instantiates a whole hypothetical game state with the
hypothetical tile placed with the idea that we can use that to simulate
and display hypothetical flows, but the UI doesn't show that yet. It
also doesn't distinguish between the hypothetical tile and currently
placed tiles, you just have to infer from the fact that one tile moves
around with your cursor.