Update app.rs to swap views in hotseat mode.
In the current in-memory mode the game is a bit annoying to play
because you constantly have to click the current player. Change
app.rs to swap views (once) automatically whenever a move is
played so that the players can just alternate taking moves on
a single computer in a "hot seat" mode.