Rewrite server backend to use Tokio.
Instead of spawning a bunch of threads and having each poll for updates
to the game, it just directly pushes any new actions to each client
after a move is successfully made in the game.

The server architecture is easy to extend into a server that handles
more than one game, just by adding more message types that the client
can send, to e.g. allow clients to subscribe/unsubscribe from multipe
games.