Fix bug with multiple websocket client connections.
There was a bug where if a websocket client had connected and then
disconnected, no other clients (websocket or native) could connect to
the server. I couldn't figure out why this was happening but
I hypothesize that it was either the disconnect causing an error in the
server which silently crashed something, or some blocking operation
(heartbeating?) that got the server stuck once the other end of the
websocket was no longer available.

I fixed the bug by switching from tokio-websockets to tokio-tungstenite,
which was an easy swap and everything just works now.