Unify client code for wasm and native clients.
The unified interface for interacting with the server is a synchronous
interface that lets you poll for all the new messages from the server
(since you last polled) or send a message. The server backend is built
on top of the connection (which is implemented for websockets and TCP)
and works the same way for both types of client. This will make it
easier to implement the lobby because it won't need to be implemented
twice.