Simplify server backend implementation
The server backend had some RwLocks around its state because the
interface I wrote for the Backend trait requires all the accessors to
take the Backend struct as immutable, but the server backend needs to
update its internal state when each accessor is called so that it can
return the most up-to-date value.

Instead, I extended the Backend trait to have an `update` function,
and we require that users of the Backend trait call that each frame to
ensure they have the most up-to-date values when they call the
accessors. Now nothing needs locking in the business-logic parts of the
implementation.