Add hacky game history saving to server.
If you set the environment variable GAME_SAVE_DIR on the server, it will
write a file with the complete game history after each move to a file
with the same name as the room code. It never reads this dir. The
purpose is to let us start building up a history of actual games for
future reference (legal move test cases, benchmarking, AI training,
etc.)